Created restart utility module
