+++ applications/harnesses/adams_lsc_policy/train_lsc_policy.py
@@ -11,6 +11,7 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 from yoke.models.policyCNNmodules import gaussian_policyCNN
 from yoke.datasets.lsc_dataset import LSC_hfield_policy_DataSet
 import yoke.torch_training_utils as tr
+from yoke.utils.restart import continuation_setup
 from yoke.utils.dataload import make_distributed_dataloader
 from yoke.utils.checkpointing import save_model_and_optimizer
 from yoke.utils.checkpointing import load_model_and_optimizer
@@ -331,7 +332,7 @@ def main(
         #############################################
         FINISHED_TRAINING = epochIDX + 1 > total_epochs
         if not FINISHED_TRAINING:
-            new_flux_file = tr.continuation_setup(
+            new_flux_file = continuation_setup(
                 new_chkpt_path,
                 studyIDX,
                 last_epoch=epochIDX,
