+++ applications/harnesses/burr_lsc_density_surrogate/train_lsc_surrogate_LRsched.py
@@ -21,6 +21,8 @@ import torch.nn as nn
 from yoke.models.surrogateCNNmodules import tCNNsurrogate
 from yoke.datasets.lsc_dataset import LSC_cntr2rho_DataSet
 import yoke.torch_training_utils as tr
+from yoke.utils.dataload import make_dataloader
+from yoke.utils.restart import continuation_setup
 from yoke.utils.checkpointing import load_model_and_optimizer_hdf5
 from yoke.utils.checkpointing import save_model_and_optimizer_hdf5
 from yoke.helpers import cli
@@ -209,10 +211,10 @@ if __name__ == "__main__":
     ending_epoch = min(starting_epoch + cycle_epochs, total_epochs + 1)
 
     # Setup Dataloaders
-    train_dataloader = tr.make_dataloader(
+    train_dataloader = make_dataloader(
         train_dataset, batch_size, train_batches, num_workers=num_workers
     )
-    val_dataloader = tr.make_dataloader(
+    val_dataloader = make_dataloader(
         val_dataset, batch_size, val_batches, num_workers=num_workers
     )
 
@@ -267,7 +269,7 @@ if __name__ == "__main__":
     #############################################
     FINISHED_TRAINING = epochIDX + 1 > total_epochs
     if not FINISHED_TRAINING:
-        new_slurm_file = tr.continuation_setup(
+        new_slurm_file = continuation_setup(
             new_h5_path, studyIDX, last_epoch=epochIDX
         )
         os.system(f"sbatch {new_slurm_file}")
