+++ applications/harnesses/ch_lightning_loderunner/train_LodeRunner_lightning.py
@@ -27,7 +27,7 @@ import numpy as np
 from yoke.models.vit.swin.bomberman import LodeRunner, Lightning_LodeRunner
 from yoke.datasets.lsc_dataset import LSCDataModule
 from yoke.datasets.transforms import ResizePadCrop
-import yoke.torch_training_utils as tr
+from yoke.utils.restart import continuation_setup
 from yoke.utils.parameters import freeze_torch_params
 from yoke.lr_schedulers import CosineWithWarmupScheduler
 from yoke.helpers import cli
@@ -269,7 +269,7 @@ if __name__ == "__main__":
     if trainer.is_global_zero:
         FINISHED_TRAINING = (final_epoch + 1) >= args.total_epochs
         if not FINISHED_TRAINING:
-            new_slurm_file = tr.continuation_setup(
+            new_slurm_file = continuation_setup(
                 checkpoint_callback.last_model_path,
                 args.studyIDX,
                 last_epoch=final_epoch + 1,
